Fixes for re-enabling more MSVC level 4 warnings: content/browser/ edition
This contains fixes for the following sorts of issues:
* Signedness mismatch
* Assignment inside conditional
* Possibly-uninitialized local variable

This also contains a very small number of other cleanups to nearby code.

BUG=81439
TEST=none

Review URL: https://codereview.chromium.org/373873002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@283169 0039d316-1c4b-4281-b951-d872f2087c98

+++ content/browser/frame_host/navigation_controller_impl.cc
@@ -47,7 +47,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 namespace content {
 namespace {
 
-const int kInvalidateAll = 0xFFFFFFFF;
+const unsigned kInvalidateAll = 0xFFFFFFFF;
 
 // Invoked when entries have been pruned, or removed. For example, if the
 // current entries are [google, digg, yahoo], with the current entry google,
@@ -148,7 +148,7 @@ bool ShouldKeepOverride(const NavigationEntry* last_entry) {
 
 // NavigationControllerImpl ----------------------------------------------------
 
-const size_t kMaxEntryCountForTestingNotSet = -1;
+const size_t kMaxEntryCountForTestingNotSet = static_cast<size_t>(-1);
 
 // static
 size_t NavigationControllerImpl::max_entry_count_for_testing_ =
